backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
pydantic-settings
alembic
